added file check before copying prompttag boiler plate

+++ install.sh
@@ -5,5 +5,6 @@ rm -rf ~/.vim
 mv vimrc ~/.vimrc
 mv bashrc ~/.bashrc
 mv vim ~/.vim;
-mv prompttag ~/.bashprompttag
-source ~/.bashrc
+if [ ! -f ~/.bashprompttag ]; then
+  mv prompttag ~/.bashprompttag
+fi
